requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-storage-blob
requests
certifi

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Additional Azure services
azure-identity>=1.12.0
azure-mgmt-datafactory>=2.0.0
azure-mgmt-storage>=20.0.0

# Data processing
pandas>=1.5.0
pyarrow>=10.0.0

# Monitoring and logging
opencensus-ext-azure>=1.1.0
opencensus-ext-logging>=0.1.0